requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrify==3.0
attrs==25.1.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.0
requests==2.32.3
tracemoepy==4.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
supabase
